+++ bindings/python/src/modelling/expose-fly-high.cpp
@@ -1,7 +1,7 @@
 /// @file
 /// @copyright Copyright (C) 2023 LAAS-CNRS, INRIA
-#include "aligator/modelling/multibody/context.hpp"
 #ifdef ALIGATOR_WITH_PINOCCHIO
+#include "aligator/modelling/multibody/context.hpp"
 
 #include "aligator/python/fwd.hpp"
 #include "aligator/python/modelling/multibody-utils.hpp"