+++ bindings/python/src/modelling/expose-gravity-compensation.cpp
@@ -1,3 +1,5 @@
+/// @copyright Copyright (C) 2024 LAAS-CNRS, INRIA
+#ifdef ALIGATOR_WITH_PINOCCHIO
 #include "aligator/python/fwd.hpp"
 #include "aligator/modelling/multibody/context.hpp"
 #include "aligator/modelling/multibody/gravity-compensation-residual.hpp"
@@ -30,3 +32,4 @@ void exposeGravityCompensation() {
 }
 
 } // namespace aligator::python
+#endif