+++ include/aligator/modelling/multibody/contact-force.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
+#ifdef ALIGATOR_PINOCCHIO_V3
+
 #include "./fwd.hpp"
 #include "aligator/core/function-abstract.hpp"
 
 #include <Eigen/src/Core/util/Constants.h>
 #include <proxsuite-nlp/modelling/spaces/multibody.hpp>
-
-#ifdef ALIGATOR_PINOCCHIO_V3
 #include <pinocchio/algorithm/proximal.hpp>
 
 namespace aligator {