+++ include/aligator/modelling/multibody/multibody-friction-cone.hxx
@@ -1,5 +1,6 @@
 #pragma once
 
+#ifdef ALIGATOR_PINOCCHIO_V3
 #include "aligator/modelling/multibody/multibody-friction-cone.hpp"
 
 #include <pinocchio/algorithm/constrained-dynamics.hpp>
@@ -84,3 +85,4 @@ MultibodyFrictionConeDataTpl<Scalar>::MultibodyFrictionConeDataTpl(
 }
 
 } // namespace aligator
+#endif