+++ include/aligator/modelling/multibody/multibody-wrench-cone.hxx
@@ -2,6 +2,7 @@
 
 #include "aligator/modelling/multibody/multibody-wrench-cone.hpp"
 
+#ifdef ALIGATOR_PINOCCHIO_V3
 #include <pinocchio/algorithm/constrained-dynamics.hpp>
 #include <pinocchio/algorithm/constrained-dynamics-derivatives.hpp>
 
@@ -65,3 +66,4 @@ MultibodyWrenchConeDataTpl<Scalar>::MultibodyWrenchConeDataTpl(
 }
 
 } // namespace aligator
+#endif