+++ src/compat/crocoddyl/problem-wrap.cpp
@@ -1,5 +1,3 @@
-/// @file    instantiate.cpp
-/// @details Instantiates the templates for the chosen context's Scalar type.
 /// @copyright Copyright (C) 2022-2024 LAAS-CNRS, INRIA
 #include "aligator/compat/crocoddyl/problem-wrap.hxx"