+++ src/modelling/multibody/contact-force.cpp
@@ -1,5 +1,5 @@
-#include "aligator/modelling/multibody/contact-force.hxx"
 #ifdef ALIGATOR_PINOCCHIO_V3
+#include "aligator/modelling/multibody/contact-force.hxx"
 namespace aligator {
 
 template struct ContactForceResidualTpl<context::Scalar>;