+++ src/modelling/multibody/multibody-friction-cone.cpp
@@ -1,5 +1,5 @@
-#include "aligator/modelling/multibody/multibody-friction-cone.hxx"
 #ifdef ALIGATOR_PINOCCHIO_V3
+#include "aligator/modelling/multibody/multibody-friction-cone.hxx"
 
 namespace aligator {